% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_gram.R
\name{lasso_gram}
\alias{lasso_gram}
\title{Solving Lasso Regression}
\usage{
lasso_gram(
  xtx,
  xty,
  lambda,
  penalty_factors,
  beta0 = NULL,
  max_iter = 1000,
  tolerance = 1e-06
)
}
\arguments{
\item{xtx}{The quadratic coefficient \eqn{A}, usually corresponding to the data gram matrix \eqn{X X' / T} in Lasso.}

\item{xty}{The linear coefficient \eqn{\alpha}, usually corresponding to \eqn{X Y' / T} in Lasso.}

\item{lambda}{The penalty coefficient.}

\item{penalty_factors}{Whether the variables is penalized, and its penalty scale with respect to \code{lambda}.}

\item{beta0}{The initialization for the estimator.}

\item{max_iter}{Maximum number of iterations.}

\item{tolerance}{Convergence criterion.}
}
\value{
The estimated sparse vector \code{beta}.
}
\description{
For a Lasso regression problem of the format
\deqn{f(\beta) = \frac{1}{2} \beta' A \beta - \beta' \alpha + \lambda \|\beta\|_1,}
this is the function to solve it using the input of \eqn{A} and \eqn{\alpha}.
}
\examples{
X = matrix(rnorm(40), 4, 10); Y = rnorm(10)
lasso_gram(tcrossprod(X), tcrossprod(X, t(Y)), 0.05, rep(1, 4))
}
