% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PVAR.FiNNO-package.R
\docType{package}
\name{PVAR.FiNNO-package}
\alias{PVAR.FiNNO}
\alias{PVAR.FiNNO-package}
\title{PVAR.FiNNO: Fixed-Nuclear-Norm Optimization algorithm for Panel Vector Auto-Regression}
\description{
An ADMM approach to solve the estimation problem of Panel Vector Auto Regression with Fixed Nuclear Norm constraints.
}
\keyword{internal}
