% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateData.R
\name{sampleData}
\alias{sampleData}
\title{Simulate PVAR time series data}
\usage{
sampleData(
  Coef,
  M = dim(Coef$A)[1],
  p = dim(Coef$A)[2],
  r = NULL,
  s = NULL,
  TT = NULL,
  n = 1
)
}
\arguments{
\item{Coef}{The PVAR coefficients, the first element of the output list generated from \code{simuPar}.}

\item{M}{Size of the panel, i.e., number of entities.}

\item{p}{Dimension of the PVAR, i.e., number of variables.}

\item{r}{Rank of the low-rank component.}

\item{s}{The average fraction of nonzero elements in the sparse components of the coefficient matrices.}

\item{TT}{The length of time series, default a scalar \code{TT = p * r * 2}. The user can input an integer or a lengths-\code{M} vector of integers. If the length of the input integer vector is less than \code{M}, the first number is recycled.}

\item{n}{The index of the replicate of the simulated time series.}
}
\value{
\code{Data} A list of panel time series data with dimension \code{nM} x \code{np} x \code{nr} x \code{ns} corresponding to the number of different parameter combinations. Each entry in the list is a length-\code{N} list of named lists\itemize{
\item \code{XTS}: A length-\code{M} list of time series, with the \code{m}-th data matrix having size \code{p} x \code{(TT_m + 1)}.
\item \code{ind}: The index of the replicate in the simulation, ranging from 1 to \code{N}.
\item \code{M, p, r, s, TT}: the same as the input, for bookkeeping purpose.
}
}
\description{
Simulate PVAR time series data
}
\examples{
sampleData(simuPar(5, 10, 3, 0.02)$Coef, 5, 10, 3, 0.02)
}
