% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateData.R
\name{simuPar}
\alias{simuPar}
\title{Simulate PVAR Parameters}
\usage{
simuPar(
  M,
  p,
  r,
  s,
  C = sqrt(p * r),
  G.W = NULL,
  G.S = NULL,
  isolate = 0,
  sg_w = 0,
  sg_s = 0,
  GW.sd = 0,
  GS.sd = 0,
  GS.frac = 0
)
}
\arguments{
\item{M}{Size of the panel, i.e., number of entities.}

\item{p}{Dimension of the PVAR, i.e., number of variables.}

\item{r}{Rank of the low-rank component.}

\item{s}{The average fraction of nonzero elements in the sparse components of the coefficient matrices.}

\item{C}{The specified nuclear norm. Default is \code{C = sqrt(p * r).}}

\item{G.W}{Number of rescaling groups in the panel, excluding singular ones and isolates. Default is \code{G.W = M - sg_w - sg_s - isolate}.}

\item{G.S}{Sparse group assignment in the panel, excluding \code{sg_s} singular ones. Default is \code{G.S = 1:(M - sg_s)}.}

\item{isolate}{Integer or fraction (\code{isolate = round(isolate * M)}), the number of rescaling isolates in the panel model.}

\item{sg_w}{Integer or fraction (\code{sg_w = round(sg_w * M)}), the number of entities with singular W (i.e., purely sparse entities) in the panel.}

\item{sg_s}{Integer or fraction (\code{sg_s = round(sg_s * M)}), the number of entities with singular S (i.e., purely low-rank entities) in the panel.}

\item{GW.sd}{The standard deviation of the rescaling effects \code{W} in the same group when \code{GW.sd >= 0}. Default is \code{GW.sd = 0} meaning exact equality in the group. If \code{GW.sd < 0}, then \code{-GW.sd} is the ratio of standard deviation versus the mean magnitude of \code{W} in the same group.}

\item{GS.sd}{The standard deviation of the sparse components \code{S} in the same group when \code{GS.sd >= 0}. Default is \code{GS.sd = 0} meaning exact equality in the group. If \code{GS.sd < 0}, then \code{-GS.sd} is the ratio of standard deviation versus the mean magnitude of \code{S} in the same group.}

\item{GS.frac}{The density ratio of the sparse perturbation versus mean in the same group. Default is \code{GS.frac = 0} meaning no extra non-zero elements.}
}
\value{
A named list:\itemize{
\item \code{Coef}: A named list of coefficients:\itemize{
\item \code{A}: An \code{M} x \code{p} x \code{p} array of \code{M} coefficient matrices of size \code{p} x \code{p}.
\item \code{Sigma}: A length-\code{M} vector of the scalar variances \eqn{\sigma_m}, and the covariance matrices are then constructed as \eqn{\Sigma_m = \sigma_m^2 I_p}.
\item \code{W}: An \code{M} x \code{p} matrix of rescaling effects, with rows corresponding to entities.
\item \code{Phi}: The shared \code{p} x \code{p} low rank basis.
\item \code{S}: A length-\code{M} list of \code{p} x \code{p} sparse matrices.
}
\item \code{r, M, p, s}: the same as the input, for bookkeeping purpose.
}
}
\description{
Aims to simulate Panel VAR data parameters, including the transition matrices of the form \eqn{A_m = W_m \Phi + S_m} with all these components, and covariance matrices of the innovations being scalar matrices \eqn{\Sigma_m = \sigma_m^2 I_p}.
}
\details{
If there exist mixture patterns in the simulated parameters, they are arranged in the following order: 1. Clusters; 2. Isolates; 3. Singular S; Singular W.
}
\examples{
data = simuPar(5, 10, 3, 0.02)
}
