% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PVAR_ADMM.R
\name{PVAR_ADMM}
\alias{PVAR_ADMM}
\title{ADMM Algorithm for Panel Vector Auto-Regression Model}
\usage{
PVAR_ADMM(
  XTS,
  r,
  eta,
  TT = sapply(XTS, ncol) - 1,
  M = length(XTS),
  p = nrow(XTS[[1]]),
  C = sqrt(p * r),
  rho = M/10,
  maxiter = 10000,
  miniter = 200,
  err = 1e-05,
  pb = NULL,
  verbose = FALSE,
  Phi_BL = NULL,
  Phi = NULL,
  Gamma = NULL,
  WS = NULL,
  bulk = 1,
  perupdate = 1,
  kappa = NULL,
  normalize = T
)
}
\arguments{
\item{XTS}{A length-\code{M} list of time series, where the \code{m}-th data matrix has size \code{p} x \code{(TT_m + 1)} with potentially unbalanced number of observations \code{TT_m} from \code{TT}.}

\item{r}{The upper bound for the rank of the estimated low-rank component.}

\item{eta}{The penalty coefficient for the sparse component.}

\item{TT}{The effective lengths of the time series, i.e., the notations \eqn{T_m} in description above.}

\item{M}{The number of subjects in the panel.}

\item{p}{The number of variables in the multivariate time series data.}

\item{C}{The specified nuclear norm for the low-rank component.}

\item{rho}{The step size coefficient in the ADMM setting. If is \code{NULL}, it will be adaptively selected in every iteration to balance the subproblem of \eqn{\Phi}.}

\item{maxiter}{Maximum number of iterations.}

\item{miniter}{Minimum number of iterations.}

\item{err}{Tolerance error to determine convergence.}

\item{pb}{The progress bar object defined from the \code{'progressr'} package to track the estimation progress. By default \code{pb = NULL}, no progress bar is inputted and progress is instead tracked by \code{message()} if \code{verbose = TRUE}.}

\item{verbose}{Logical, whether verbal messages should be printed for progress tracking. Effective only when \code{pb} is \code{NULL}.}

\item{Phi_BL}{The initializer for the augmented variable \eqn{\Phi_c}.}

\item{Phi}{The initializer for \eqn{\Phi}.}

\item{Gamma}{The initializer for the dual variable (Lagrange multiplier) \eqn{\Gamma}.}

\item{WS}{The list of initializers for the rescaling \code{W} (a matrix of size \code{M} x \code{p}) and the sparse components \code{S} (a length-\code{M} list of \code{p} x \code{p} matrices).}

\item{bulk}{If \code{verbose = TRUE}, the number of iterations between permanent progress tracking messages.}

\item{perupdate}{If \code{verbose = TRUE}, the number of iterations between live updates about progress tracking.}

\item{kappa}{The proximal step size coefficient for the subproblem of \eqn{\Phi_c}, i.e., \code{Phi_BL}.}

\item{normalize}{Logical, whether the time series data should be normalized. Default is True.}
}
\value{
A named list of estimators and some metrics:\itemize{
\item \code{Phi}: the \code{p} x \code{p} estimator of \eqn{\Phi} in the objective function above, not necessarily low-rank.
\item \code{W}: the \code{M} x \code{p} estimated rescaling effects, with each row corresponding to \eqn{W_m}.
\item \code{S}: a length-\code{M} list of sparse components, with the entries being \eqn{S_m} of dimension \code{p} x \code{p}.
\item \code{Phi_BL}: the \code{p} x \code{p} estimator of \eqn{\Phi_c} in the objective function above, that is strictly constrained with low-rankness.
\item \code{Gamma}: the \code{p} x \code{p} estimator of the dual variable matrix \eqn{\Gamma} in the objective function above.
\item \code{status}: 1 if convergence criteria is met within \code{maxiter} iterations, otherwise 0.
\item \code{iters}: if \code{status == 1}, it is the number of iterations until reaching convergence; otherwise \code{maxiter} (when \code{status == 0}).
\item \code{traj}: a vector of length \code{iters}, the trajectory of the evaluation of the objective function along the sequence of estimators.
\item \code{rele}: a vector of length \code{iters}, the trajectory of the relative errors of the estimator \eqn{\Phi} between consecutive iterations.
\item \code{ics}: the vector of RSS/AIC/BIC/HQC/eBIC evaluation at the final output estimators.
\item \code{time}: computation time for the whole estimation process.
\item \code{eta, C, rho}: the same as the input, for bookkeeping purpose.
}
}
\description{
Consider a multivariate Panel VAR model of the form
\deqn{X_t^m = A_m X_{t-1}^m + \epsilon_t^m}
with the coefficient matrices \eqn{A_m} have the structure
\deqn{A_m = W_m \Phi + S_m,}
where \eqn{W_m} are diagonal matrices, \eqn{S_m} are sparse matrices, and \eqn{\Phi}
is the low-rank basis shared by the whole panel. The optimization problem is constructed under the ADMM framework, with the objective function
\deqn{G(W, S, \Phi, \Phi_c, \Gamma; \eta, \rho) = \sum_{m=1}^M (\frac{1}{2T_m} \|Y_m - (W_m \Phi + S_m) X_m\|_F^2 + \eta \|S_m\|_1)\\ + \rho \|\Phi - \Phi_c\|_F^2 + \rho \langle \Gamma, \Phi - \Phi_c \rangle}
subject to fixed nuclear-norm constraint \eqn{\|\Phi_c\|_* = \ell} and row-balanced constraint \eqn{\|e_i' \Phi_c\| = \|e_j' \Phi_c\|} for any \eqn{i \ne j}. Note here \eqn{Y_m = (X_1^m, \dots, X_T^m)} and \eqn{X_m = (X_0^m, \dots, X_{T-1}^m)} are stacked time series data of dimension \eqn{p \times T_m}.
}
\details{
While constructing the objective function, note that \eqn{\Phi} is set to be flexible, and all the low-rank constraints are imposed through specifying the support domain of the augmented variable \eqn{\Phi_c}. It may be shown that the optimization within the fixed nuclear norm space can automatically induce sparsity in the singular values, as a simplex vertex of the support domain is likely to be hit. For more details, see (CITE).
}
\examples{
DP = simuDP(1, M = 5, p = 10, r = 3, s = 0.02)
est = PVAR_ADMM(DP$Data$XTS, 5, 0.01, maxiter = 10, miniter = 3)
}
