% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateData.R
\name{simuDP}
\alias{simuDP}
\title{Simulate PVAR parameters and time series data}
\usage{
simuDP(
  N,
  TT = NULL,
  Pars = NULL,
  M = NULL,
  p = NULL,
  r = NULL,
  s = NULL,
  C = 1,
  G = NULL,
  outlier = 0,
  sg_w = 0,
  sg_s = 0,
  seed = NULL,
  prl = NULL
)
}
\arguments{
\item{N}{Number of replicates of simulated time series per setting.}

\item{TT}{The length of time series, default \code{TT = p * r * 2}.}

\item{Pars}{The PVAR parameters generated from \code{simuPar}.}

\item{M}{Size of the panel, i.e., number of entities; can be a vector.}

\item{p}{Dimension of the PVAR, i.e., number of variables; can be a vector.}

\item{r}{Rank of the low-rank component; can be a vector.}

\item{s}{The average fraction of nonzero elements in the sparse components of the coefficient matrices; can be a vector}

\item{C}{The specified nuclear norm.}

\item{G}{Number of groups in the panel.}

\item{outlier}{Integer or fraction, the size of isolates in the panel model.}

\item{sg_w}{Integer or fraction (\code{sg_w = round(sg_w * M)}), the size of entities with singular W (i.e., purely sparse entities) in the panel.}

\item{sg_s}{Integer or fraction (\code{sg_s = round(sg_s * M)}), the size of entities with singular S (i.e., purely low-rank entities) in the panel.}

\item{seed}{Random seed.}

\item{prl}{If \code{is.numeric(prl)} and \code{prl >= 1}, then its rounded integer is treated as the number of cores for parallel simulation. By default \code{prl = NULL} and simulations are generated sequentially.}
}
\value{
A named list of PVAR parameters and data. In particular denote \code{nM = length(M)}, \code{np = length(p)}, \code{nr = length(r)} and \code{ns = length(s)}, then\itemize{
\item \code{Data} A list of panel time series data with dimension \code{nM} x \code{np} x \code{nr} x \code{ns} corresponding to the number of different parameter combinations. Each entry in the list is a length-\code{N} list of named lists\itemize{
\item \code{XTS}: \code{M} x \code{p} x \code{(TT + 1)} panel time series.
\item \code{ind}: The index of the replicate in the simulation, ranging from 1 to \code{N}.
\item \code{M, p, r, s, TT}: the same as the input, for bookkeeping purpose.
}
\item \code{Pars} is a list of simulated PVAR coefficients with the same dimension as \code{Data}, corresponding to all the combinations of the parameters. Each entry is an output object returned by the function \code{simuPar}.
}
}
\description{
One can either input all the coefficients \code{(M, p, r, s)} to start from sampling the coefficient matrices, or it also supports inputting sampled coefficient matrices from \code{simuPar} for only time series simulation. In addition, if the input parameters \code{(M, p, r, s)} are of the vector format, then every combination of the setup is simulated.
}
\examples{
simuDP(1, M = 5, p = 10, r = 3, s = 0.02, prl = 2)
}
