% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateData.R
\name{simuPar}
\alias{simuPar}
\title{Simulate PVAR Parameters}
\usage{
simuPar(M, p, r, s, C = 1, G = NULL, isolate = 0, sg_w = 0, sg_s = 0, G.sd = 0)
}
\arguments{
\item{M}{Size of the panel, i.e., number of entities.}

\item{p}{Dimension of the PVAR, i.e., number of variables.}

\item{r}{Rank of the low-rank component.}

\item{s}{The average fraction of nonzero elements in the sparse components of the coefficient matrices.}

\item{C}{The specified nuclear norm.}

\item{G}{Number of groups in the panel.}

\item{isolate}{Integer or fraction (\code{isolate = round(isolate * M)}), the number of isolates in the panel model.}

\item{sg_w}{Integer or fraction (\code{sg_w = round(sg_w * M)}), the number of entities with singular W (i.e., purely sparse entities) in the panel.}

\item{sg_s}{Integer or fraction (\code{sg_s = round(sg_s * M)}), the number of entities with singular S (i.e., purely low-rank entities) in the panel.}

\item{G.sd}{The ratio of standard deviation versus the norm of the rescaling effects \code{W} in the same group. Default is \code{G.sd = 0} meaning exact equality in the group.}
}
\value{
A named list:\itemize{
\item \code{Coef}: A named list of coefficients:\itemize{
\item \code{A}: An \code{M} x \code{p} x \code{p} array of \code{M} coefficient matrices of size \code{p} x \code{p}.
\item \code{Sigma}: A length-\code{M} vector of the scalar variances \eqn{\sigma_m}, and the covariance matrices are then constructed as \eqn{\Sigma_m = \sigma_m^2 I_p}.
\item \code{W}: An \code{M} x \code{p} matrix of rescaling effects, with rows corresponding to entities.
\item \code{Phi}: The shared \code{p} x \code{p} low rank basis.
\item \code{S}: A length-\code{M} list of \code{p} x \code{p} sparse matrices.
}
\item \code{r, M, p, s}: the same as the input, for bookkeeping purpose.
}
}
\description{
Aims to simulate Panel VAR data parameters, including the transition matrices of the form \eqn{A_m = W_m \Phi + S_m} with all these components, and covariance matrices of the innovations being scalar matrices \eqn{\Sigma_m = \sigma_m^2 I_p}.
}
\details{
If there exist mixture patterns in the simulated parameters, they are arranged in the following order: 1. Clusters; 2. Isolates; 3. Singular S; Singular W.
}
\examples{
data = simuPar(5, 10, 3, 0.02)
}
